[FLINK-25810][connector/kinesis] Renaming e2e test module for kinesis data streams.
