+++ flink-connectors/flink-sql-connector-aws-kinesis-data-streams/src/main/resources/META-INF/NOTICE
@@ -43,6 +43,7 @@ This project bundles the following dependencies under the Apache Software Licens
 - com.typesafe.netty:netty-reactive-streams:2.0.5
 - org.apache.httpcomponents:httpclient:4.5.13
 - org.apache.httpcomponents:httpcore:4.4.14
+- commons-logging:commons-logging:1.1.3
 
 
 This project bundles the following dependencies under the Creative Commons Zero license (https://creativecommons.org/publicdomain/zero/1.0/).
