+++ flink-end-to-end-tests/pom.xml
@@ -80,7 +80,7 @@ under the License.
 		<module>flink-glue-schema-registry-avro-test</module>
 		<module>flink-glue-schema-registry-json-test</module>
 		<module>flink-end-to-end-tests-scala</module>
-		<module>flink-end-to-end-tests-kinesis-streams</module>
+		<module>flink-end-to-end-tests-aws-kinesis-data-streams</module>
 		<module>flink-end-to-end-tests-aws-kinesis-firehose</module>
 	</modules>
